pyproject.toml
[tool.poetry]
name = "wikiracer"
version = "0.1.0"
description = ""
authors = ["Ben Sandler <ben.sandler@opendoor.com>"]

[tool.poetry.dependencies]
python = "^3.7"
black = "^19.10b0"
pymediawiki = "^0.6.3"
mypy = "^0.770"
flake8 = "^3.7.9"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
